requirements.txt
streamlit
transformers
sentencepiece
gTTS
openai-whisper
yt-dlp
fpdf
vaderSentiment
yake
torch
PyPDF2
pymupdf
arxiv
requests
beautifulsoup4
nltk
spacy
